removed custom batch norm 1d
